+++ src/lib.rs
@@ -2,7 +2,6 @@
 
 // Public modules
 mod common;
-mod error;
 
 #[cfg(feature = "async")]
 pub mod async_dht;
@@ -13,6 +12,6 @@ pub mod server;
 pub use crate::common::{Id, MutableItem};
 pub use bytes::Bytes;
 pub use dht::{Dht, Testnet};
+pub use rpc::PutError;
 
 pub use ed25519_dalek::SigningKey;
-pub use error::{Error, Result};